Update Check Time Boundaries.yaml - Boolean values do not work in "Condition" Action
Testing this action natively, the boolean value that is associated with the "outside_working_hours" field are unusable within a "Condition" action (even when utilizing the boolean == and bool value for true/false". 

When switching the field from text to string and making the values set as "false" & "true", this does work within a Condition action.

I am proposing this PR, however I am open to other alternative options utilizing the native action.