Signal-Filter-Fields.Yaml
The purpose of this custom action is to deconstruct a CSE Insight based on an assigned filter in order to retrieve specific field elements from the Signal Array that match the criteria. Currently there is no way (outside of a custom action) to prevent a playbook from executing against multiple matches and instead to only focus on the elements that match the condition. 

Ex Use Case:

Insight A contained 3 different signals with 3 different device_ip addresses. This action will allow you to target explicitly the signal that matches your filter and only return the device_ip that matches your condition.